feat(menu): add image support to menu items and update navbar display

+++ web/app/Models/Setting.php
@@ -122,7 +122,9 @@ class Setting extends Model
                 // if the image was erased
                 if ($value == null) {
                     // delete the image from disk
-                    removeFile($this->{$attributeName}, $disk);
+                    if (!empty($this->{$attributeName})) {
+                        removeFile($this->{$attributeName}, $disk);
+                    }
 
                     // set null in the database column
                     $this->attributes[$attributeName] = null;
@@ -134,7 +136,9 @@ class Setting extends Model
                     $filename = $filename . '.svg';
                     $value = str_replace('data:image/svg+xml;base64,', '', $value);
                     $value = str_replace(' ', '+', $value);
-                    removeFile($this->{$attributeName}, $disk);
+                    if (!empty($this->{$attributeName})) {
+                        removeFile($this->{$attributeName}, $disk);
+                    }
                     if (!File::exists($destinationPath)) {
                         mkdir($destinationPath);
                     }
@@ -149,7 +153,9 @@ class Setting extends Model
                     // 2. Store the image on disk.
                     //Optimize max size and save
                     resizeImage($image, null, null);
-                    removeFile($this->{$attributeName}, $disk);
+                    if (!empty($this->{$attributeName})) {
+                        removeFile($this->{$attributeName}, $disk);
+                    }
                     saveImage($disk, $destinationPath . '/' . $filename, $image);
 
                     // 3. Save the path to the database
